Filter out disputes for status queries
