+++ index-handlers/src/main/java/no/sikt/nva/nvi/index/utils/AggregationFunctions.java
@@ -1,56 +1,17 @@
 package no.sikt.nva.nvi.index.utils;
 
-import static java.util.Objects.nonNull;
-import java.util.Arrays;
 import java.util.Map;
-import org.opensearch.client.json.JsonData;
-import org.opensearch.client.opensearch._types.FieldValue;
 import org.opensearch.client.opensearch._types.aggregations.Aggregation;
 import org.opensearch.client.opensearch._types.aggregations.NestedAggregation;
 import org.opensearch.client.opensearch._types.aggregations.SumAggregation;
 import org.opensearch.client.opensearch._types.aggregations.TermsAggregation;
-import org.opensearch.client.opensearch._types.query_dsl.BoolQuery.Builder;
-import org.opensearch.client.opensearch._types.query_dsl.MatchAllQuery;
-import org.opensearch.client.opensearch._types.query_dsl.MatchQuery;
-import org.opensearch.client.opensearch._types.query_dsl.NestedQuery;
 import org.opensearch.client.opensearch._types.query_dsl.Query;
-import org.opensearch.client.opensearch._types.query_dsl.RangeQuery;
-import org.opensearch.client.opensearch._types.query_dsl.TermQuery;
 
 public final class AggregationFunctions {
 
     private static final CharSequence DELIMITER = ".";
 
     private AggregationFunctions() {
-    }
-
-    public static Query nestedQuery(String path, Query query) {
-        return new NestedQuery.Builder()
-                   .path(path)
-                   .query(query)
-                   .build()._toQuery();
-    }
-
-    public static Query mustMatch(Query... queries) {
-        return new Query.Builder()
-                   .bool(new Builder().must(Arrays.stream(queries).toList()).build())
-                   .build();
-    }
-
-    public static Query fieldValueQuery(String field, String value) {
-        return nonNull(value) ? new TermQuery.Builder()
-                                    .field(field)
-                                    .value(getFieldValue(value))
-                                    .build()._toQuery()
-                   : matchAllQuery();
-    }
-
-    public static Query rangeFromQuery(String field, int greaterThanOrEqualTo) {
-        return new RangeQuery.Builder().field(field).gte(JsonData.of(greaterThanOrEqualTo)).build()._toQuery();
-    }
-
-    public static String joinWithDelimiter(String... args) {
-        return String.join(DELIMITER, args);
     }
 
     public static Aggregation filterAggregation(Query filterQuery) {
@@ -72,16 +33,8 @@ public final class AggregationFunctions {
                    .build();
     }
 
-    public static Query mustNotMatch(String value, String field) {
-        return new Builder()
-                   .mustNot(matchQuery(value, field))
-                   .build()._toQuery();
-    }
-
-    public static Query matchQuery(String value, String field) {
-        return new MatchQuery.Builder().field(field)
-                   .query(getFieldValue(value))
-                   .build()._toQuery();
+    public static String joinWithDelimiter(String... args) {
+        return String.join(DELIMITER, args);
     }
 
     public static NestedAggregation nestedAggregation(String... paths) {
@@ -92,13 +45,5 @@ public final class AggregationFunctions {
 
     public static Aggregation sumAggregation(String... paths) {
         return new SumAggregation.Builder().field(joinWithDelimiter(paths)).build()._toAggregation();
-    }
-
-    private static Query matchAllQuery() {
-        return new MatchAllQuery.Builder().build()._toQuery();
-    }
-
-    private static FieldValue getFieldValue(String value) {
-        return new FieldValue.Builder().stringValue(value).build();
     }
 }
